ddl: rename test name and clean up
